Updated project to current development state.
Updated the files to where I am on the development. 4 of 5 of the planned features have been added and some of the following versions as well. Just because I'm very bad at planning the order in which I'll release features ^^. All that's left to to add for v0.2-a2 is melee weapons. The skeleton of the MeleeWeapon class has already been added but not implemented. The build doesn't work anymore because I just changed the map loading handler and didn't fix one of the bugs.